refactor(view): P6-A3 second cut — extract design_codegen.cpp
Phase 6 A3 (design_import.cpp split), second cut — stacked on the
design_tokens.cpp extraction. The IR → code-generation surface
relocated into core/view/src/design_codegen.cpp (859 lines):

  * web-compat JS generator   — @pulp/react-compatible JS (default
    import target)
  * native Pulp API generator — direct Pulp widget API calls
  * generate_pulp_js()        — public dispatch over CodeGenOptions
  * js_single_quote_escape    — pulp #81 JS string-literal escaper;
    moved here since the codegen backends are its only callers

New private header core/view/src/design_import_internal.hpp declares
design_source_vendor_key() — the one design_import.cpp-defined helper
the codegen TU still calls (Phase 9 motion-provenance vendor key). The
header is scoped to core/view/src/, not the public include tree.

design_import.cpp: 4,155 → 3,324 (-831). Combined with the first cut
(design_tokens.cpp), design_import.cpp is down 4,691 → 3,324 (-29%).

Verified: pulp-view builds clean; all 3 design-import test suites pass
(pulp-test-design-import 30 cases, -react-runtime 21, -w3c-tokens 12).
